+++ silva/src/components/workflow/state.rs
@@ -171,9 +171,7 @@ impl State {
                     self.params_editor_state = Some(state);
                     self.show_params_popup = true;
                 }
-                Err(_e) => {
-                    return;
-                }
+                Err(_e) => {}
             }
         }
     }