+++ silva/src/lib.rs
@@ -8,9 +8,13 @@ pub mod components;
 pub mod headless;
 mod layout;
 mod style;
+pub mod update;
 
-pub async fn run_app<B: ratatui::backend::Backend>(terminal: &mut Terminal<B>) -> io::Result<()> {
-    let mut app = app::App::new().await;
+pub async fn run_app<B: ratatui::backend::Backend>(
+    terminal: &mut Terminal<B>,
+    update_available: Option<String>,
+) -> io::Result<()> {
+    let mut app = app::App::new(update_available).await;
     app.health_check_state.run_health_checks();
 
     loop {